search: annotate 'intentful' queries, so we can provide better results
The idea here is that as you type, we debounce the query input for 50ms
(to reduce outbound requests) and provide as-you-type fast results. Then,
once you stop typing for 500ms that is an 'intentful' query, not an autocomplete=true
one. This gives the server a chance to do more work to find better results,
e.g. digging deeper for more relevant results.

This also gives the server a chance to log queries for human review, so
we can see if people are getting good results or not, if there are errors,
etc. (just on doctree.org, and not associated with any user/IP.)

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>