frontend: download individual docs pages, not entire project
When viewing a documentation page today, e.g. `net/http` in the Go
standard library, all documentation pages for the entire repository is
downloaded (a ~7 MiB JSON file.) This is a bit too large for the
frontend to handle gracefully, and it's needless to send that much data.
The new `/api/get-page` route gives us a single page, and so we now use
that, resulting in just a few KiB JSON file download.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>