frontend: begin removing elm-spa
elm-spa has kind of been a nightmare to use, most recently I have been
unable to appropriately handle URL query param changes (when a link is
clicked to the same page with a different `?id` parameter, I want to
handle that and scroll to that position on the page) because `Effect.fromCmd`
in this context just.. does nothing. It doesn't run. I spent over 10 hours
trying to debug it, but came up dry. The page's `init` function does in
fact have that effect returned, but the task just.. never runs. There is
no error from the task, it only sends a message. The message never gets
sent.

It's also been annoying in other ways: we cannot have truly dynamic URLs
which we need for handling `/github.com/foo/bar` project routes that are
not pre-defined and so we must instead use the NotFound page to serve our
actual primary interace. It's dev server misunderstands `/github.com` to
be a file because it has 'an extension' and so it tries to read/serve that
file from disk rather than our SPA app.

Overall, now that I know a bit more about Elm I would not choose it for
most projects. I see what it's trying to do, and certainly there are use
cases it's better suited for than this - but in both projects I've done
in Elm now (codestat.dev and doctree) I have regretted using it generally.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>